Fix bin storage with `double` data type
Before this fix, the load used the `Double` object instead of `double`
Java data type to parse and store double values into Aerospike. Since `Double`
as a Java object is considered a binary, Aerospike treated the it as a `blob` and
not a `double` value which affected the type of the stored bin in Aerospike (blob instead of double).